analysis-service/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
et_xmlfile==2.0.0
fastapi==0.121.2
h11==0.16.0
idna==3.11
openpyxl==3.1.5
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pypdfium2==5.0.0
sniffio==1.3.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
pytest
httpx
pytest-asyncio
google-cloud-firestore>=2.16.0
google-auth>=2.37.0
python-multipart
